modules/vpn/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-jose[cryptography]==3.3.0
httpx==0.27.2
